# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lumière Photography Studio SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type LoginResultRole = typeof LoginResultRole[keyof typeof LoginResultRole];


export const LoginResultRole = {
  ADMIN: 'ADMIN',
  STUDIO: 'STUDIO',
} as const;

export type LoginResultStudio = {
  id?: string;
  name?: string;
};

export interface LoginResult {
  success: boolean;
  role: LoginResultRole;
  studio?: LoginResultStudio;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  ADMIN: 'ADMIN',
  STUDIO: 'STUDIO',
} as const;

export interface AuthUser {
  id: string;
  email: string;
  role: AuthUserRole;
  status?: string;
}

export interface AuthStudio {
  id: string;
  name: string;
  email: string;
  googleDriveConnected: boolean;
  /** @nullable */
  rootFolderId?: string | null;
}

export interface MeResult {
  user?: AuthUser;
  studio?: AuthStudio;
}

export interface GoogleAuthUrl {
  url: string;
}

export interface AdminStats {
  totalStudios: number;
  pendingStudios: number;
  approvedStudios: number;
  totalAlbums: number;
  totalPhotos: number;
  totalSelections: number;
}

export type StudioStatus = typeof StudioStatus[keyof typeof StudioStatus];


export const StudioStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  DISABLED: 'DISABLED',
} as const;

export interface Studio {
  id: string;
  name: string;
  email: string;
  status: StudioStatus;
  googleDriveConnected?: boolean;
  createdAt: string;
  updatedAt: string;
  albumCount?: number;
}

export interface StudiosListResult {
  studios: Studio[];
}

export interface StudioResult {
  studio: Studio;
}

export type StudioStatusUpdateStatus = typeof StudioStatusUpdateStatus[keyof typeof StudioStatusUpdateStatus];


export const StudioStatusUpdateStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  DISABLED: 'DISABLED',
} as const;

export interface StudioStatusUpdate {
  status: StudioStatusUpdateStatus;
}

export interface StudioSettingsUpdate {
  name?: string;
  password?: string;
}

export interface Album {
  id: string;
  studioId: string;
  title: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  driveFolderId?: string | null;
  allowDownload: boolean;
  allowNotes: boolean;
  maxSelection: number;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
  photoCount?: number;
  selectionCount?: number;
}

export interface AlbumInput {
  title: string;
  description?: string;
  maxSelection?: number;
  allowDownload?: boolean;
  allowNotes?: boolean;
  isPublic?: boolean;
}

export interface AlbumUpdate {
  title?: string;
  description?: string;
  maxSelection?: number;
  allowDownload?: boolean;
  allowNotes?: boolean;
  isPublic?: boolean;
}

export interface AlbumsListResult {
  albums: Album[];
}

export interface AlbumResult {
  album: Album;
}

export interface Photo {
  id: string;
  albumId: string;
  driveFileId: string;
  filename: string;
  mimeType: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  order: number;
  createdAt: string;
}

export interface AlbumDetail {
  id: string;
  studioId: string;
  title: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  driveFolderId?: string | null;
  allowDownload: boolean;
  allowNotes: boolean;
  maxSelection: number;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
  photos: Photo[];
  selectionCount?: number;
}

export interface AlbumDetailResult {
  album: AlbumDetail;
}

export interface PhotosListResult {
  photos: Photo[];
}

export interface PhotoResult {
  photo: Photo;
}

export interface Selection {
  id: string;
  albumId: string;
  photoId: string;
  customerName: string;
  /** @nullable */
  note?: string | null;
  selected: boolean;
  createdAt: string;
  updatedAt: string;
  photo?: Photo;
}

export interface SelectionsListResult {
  selections: Selection[];
}

export type PublicAlbumStudio = {
  name?: string;
};

export interface PublicAlbum {
  id: string;
  title: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  allowDownload: boolean;
  allowNotes: boolean;
  maxSelection: number;
  studio?: PublicAlbumStudio;
  photos: Photo[];
}

export interface PhotoSelectionInput {
  photoId: string;
  customerName: string;
  note?: string;
  selected: boolean;
}

export type ListAdminStudiosParams = {
status?: ListAdminStudiosStatus;
};

export type ListAdminStudiosStatus = typeof ListAdminStudiosStatus[keyof typeof ListAdminStudiosStatus];


export const ListAdminStudiosStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  DISABLED: 'DISABLED',
} as const;