# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lumière Photography Studio SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new studio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/studios:
    get:
      operationId: listAdminStudios
      tags: [admin]
      summary: List all studios
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, DISABLED]
      responses:
        "200":
          description: Studios list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudiosListResult"
  /admin/studios/{id}:
    patch:
      operationId: updateStudioStatus
      tags: [admin]
      summary: Update studio status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioResult"
    delete:
      operationId: deleteStudio
      tags: [admin]
      summary: Delete studio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /albums/{id}/photos:
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List album photos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotosListResult"
  /albums/{id}/photos/{photoId}:
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /albums/{id}/selections:
    get:
      operationId: listSelections
      tags: [albums]
      summary: List album selections
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Selections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectionsListResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    LoginResult:
      type: object
      required: [success, role]
      properties:
        success:
          type: boolean
        role:
          type: string
          enum: [ADMIN, STUDIO]
        studio:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    MeResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        studio:
          $ref: "#/components/schemas/AuthStudio"
    AuthUser:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, STUDIO]
        status:
          type: string
    AuthStudio:
      type: object
      required: [id, name, email, googleDriveConnected]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        googleDriveConnected:
          type: boolean
        rootFolderId:
          type: ["string", "null"]
...
    Studio:
      type: object
      required: [id, name, email, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, DISABLED]
        googleDriveConnected:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        albumCount:
          type: integer
    StudiosListResult:
      type: object
      required: [studios]
      properties:
        studios:
          type: array
          items:
            $ref: "#/components/schemas/Studio"
    StudioResult:
      type: object
      required: [studio]
      properties:
        studio:
          $ref: "#/components/schemas/Studio"
    StudioStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [PENDING, APPROVED, DISABLED]
...
    Photo:
      type: object
      required: [id, albumId, driveFileId, filename, mimeType, order, createdAt]
      properties:
        id:
          type: string
        albumId:
          type: string
        driveFileId:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        order:
          type: integer
        createdAt:
          type: string
    PhotosListResult:
      type: object
      required: [photos]
      properties:
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
...
    Selection:
      type: object
      required: [id, albumId, photoId, customerName, selected, createdAt, updatedAt]
      properties:
        id:
          type: string
        albumId:
          type: string
        photoId:
          type: string
        customerName:
          type: string
        note:
          type: ["string", "null"]
        selected:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        photo:
          $ref: "#/components/schemas/Photo"
    SelectionsListResult:
      type: object
      required: [selections]
      properties:
        selections:
          type: array
          items:
            $ref: "#/components/schemas/Selection"
...