# lib/api-zod/src/generated/types/meResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lumière Photography Studio SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthStudio } from './authStudio';
import type { AuthUser } from './authUser';

export interface MeResult {
  user?: AuthUser;
  studio?: AuthStudio;
}