# lib/api-zod/src/generated/types/photosListResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lumière Photography Studio SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { Photo } from './photo';

export interface PhotosListResult {
  photos: Photo[];
}